refactor(Hero): remove Hero.jsx from "Pages" folder and add it to "components" folder. refactor(Home): put Home.jsx directly in "Pages" folder. style(landingPageStyels.css): remove the file and add the style classes in the component.
