+++ src/Pages/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Hero from './Hero'
+import Hero from '@/components/Hero'
 
 const Home = () => {
   return (
