--- src/Pages/Landing page/landingPageStyels.css
@@ -1,10 +1,0 @@
-.Hero {
-    background-color: #F2F0F1;
-}
-
-.midle-num{
-    border: solid #0000001A;
-    border-top: 0;
-    border-bottom: 0;
-    padding: 15px;
-}
